Backend/requirements.txt
scikit-learn==1.2.2
pandas
numpy
seaborn
matplotlib
scipy
tabulate
joblib
python-dotenv
pyodbc
fastapi
uvicorn
xgboost
faker